Get everything working properly. Seperate dependencies and db setup.

+++ scripts/local-mac.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
+# Install script for a native Mac environment. 
+# Run this from the lightdash root directory.
 
 set -e
 
@@ -12,10 +14,8 @@ fi
 
 brew update
 
-# 2 Install nvm (https://github.com/nvm-sh/nvm#troubleshooting-on-macos) and other required dependencies
-
+# 2 Install required dependencies
 packages=(
-  "nvm"
   "pkg-config"
   "cairo"
   "pango"
@@ -26,7 +26,43 @@ packages=(
   "pixman"
   "python-setuptools"
   "postgresql@14"
+  "pgvector"
 )
+
+# 2.1
+# The Homebrew pgvector only works with posgresql@14 and postgresql@17
+# If you need a different version of postgres, you can install pgvector manually.
+# Be sure to remove it from the list of packages to install above.
+# pgvector is an extension for postgres we use in Lightdash, it needs to be installed separately
+# More info about this extension and a detailed installation guide available here: https://github.com/pgvector/pgvector
+# on Linux, you can install `postgresql-14-pgvector`, available on apt
+# You might need to point pgvector to a correct postgres instance if you have multiple versions installed
+# export PG_CONFIG=/opt/homebrew/opt/postgresql@14/bin/pg_config
+
+# **Uncomment this to install pgvector manually**
+# git clone --branch v0.8.0 https://github.com/pgvector/pgvector.git && cd pgvector && make && sudo make install && cd ..
+
+# 2.2
+# We conditionally add nvm to the list of packages to install
+# so as not to have Homebrew's collide with a manually installed nvm.
+# https://github.com/nvm-sh/nvm#troubleshooting-on-macos
+if [ -f "$HOME/.nvm/nvm.sh" ]; then
+  echo "✅ Using the existing NVM install"
+  source $HOME/.nvm/nvm.sh
+else
+  packages+=("nvm") 
+fi
+
+# Ask for confirmation before installing packages
+echo "The following packages will be installed:"
+printf '%s\n' "${packages[@]}"
+read -p "Do you want to continue? (y/N) " -n 1 -r
+echo
+if [[ ! $REPLY =~ ^[Yy]$ ]]; then
+    echo "Installation cancelled"
+    exit 1
+fi
+
 
 brew install "${packages[@]}"
 
@@ -36,13 +72,6 @@ nvm alias default v20.8.0
 
 # 4 Install postgres (https://wiki.postgresql.org/wiki/Homebrew) and pgvector
 brew services start postgresql@14
-
-# pgvector is an extension for postgres we use in Lightdash, it needs to be installed separately
-# More info about this extension and a detailed installation guide available here: https://github.com/pgvector/pgvector
-# on Linux, you can install `postgresql-14-pgvector`, available on apt
-# You might need to point pgvector to a correct postgres instance if you have multiple versions installed
-# export PG_CONFIG=/opt/homebrew/opt/postgresql@14/bin/pg_config
-git clone --branch v0.8.0 https://github.com/pgvector/pgvector.git && cd pgvector && make && sudo make install && cd ..
 
 # 5 Install dbt using pip
 # Detailed installation guide available here: https://docs.getdbt.com/docs/core/pip-install
@@ -60,17 +89,10 @@ if ! python -c "import distutils.util" 2>/dev/null; then
 fi
 python -m pip install dbt-postgres==1.4.9
 
-# 6 Clone the repo and open it in your IDE
-git clone https://github.com/lightdash/lightdash.git
-cd lightdash
-
 # 7 Copy `.env.development` to `.env.development.local`
 cp .env.development .env.development.local
 
-# 8 Edit some environment variables to match your setup
-open .env.development.local -t
-
-# 8.1 Update environment variables in .env.development.local
+# 8 Update environment variables in .env.development.local
 # You may need to edit the following variables:
 # PGUSER=pg_user *OR* machine username if no prior postgres set up
 # PGPASSWORD=pg_password *OR* blank if no prior postgres set up
@@ -87,7 +109,4 @@ sed -i '' \
 # 9 Install packages
 pnpm install
 
-# 10 Build / migrate / seed
-pnpm local-init
-
-echo "🚀 Lightdash is ready to run! Run `pnpm load:env pnpm dev` to start the app."
+echo "🚀 Dependencies are instsalled! Run `pnpm load:env pnpm local-init` to setup the databases.
